--- src/components/icons.tsx
@@ -1,13 +1,0 @@
-import { FaGithub, FaTwitter, FaYoutube } from 'react-icons/fa';
-
-export const GithubIcon: React.VFC = () => {
-  return <FaGithub />;
-};
-
-export const TwitterIcon: React.VFC = () => {
-  return <FaTwitter />;
-};
-
-export const YoutubeIcon: React.VFC = () => {
-  return <FaYoutube />;
-};
