+++ src/pages/posts/[id].tsx
@@ -1,8 +1,7 @@
 import { GetStaticPropsContext, InferGetServerSidePropsType } from 'next';
-import Head from 'next/head';
-import { Date } from 'components/date';
-import { Layout } from 'components/layout';
-import Seo from 'components/seo';
+import { Date } from 'components/Date';
+import { Layout } from 'components/Layout';
+import Seo from 'components/Seo';
 import { getAllPostIds, getPostData } from 'lib/posts';
 
 export async function getStaticPaths() {
@@ -27,10 +26,12 @@ export default function Post({ postData }: InferGetServerSidePropsType<typeof ge
   return (
     <Layout>
       <Seo pageTitle={postData.title} pageImg={postData.image} />
-      <article>
-        <h1 className='mb-8 mt-5'>{postData.title}</h1>
-        <div className='mb-16 sm:mb-20 text-sm text-center text-lightGray'>
-          <Date dateString={postData.date} />
+      <article className='w-9/12 m-auto'>
+        <div className='my-24'>
+          <h1 className='mb-6'>{postData.title}</h1>
+          <div className='text-sm text-lightGray pl-2'>
+            <Date dateString={postData.date} />
+          </div>
         </div>
         <div className='markdown' dangerouslySetInnerHTML={{ __html: postData.contentHtml }} />
       </article>
